# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API ADM inovatec-jp",
    "description": "API REST para o sistema administrativo.\n\nStatus:\n\n- **/accounts** ➜ MAPEADO e DOCUMENTADO\n- **/user** ➜ MAPEADO e DOCUMENTADO\n- **/grant** ➜ MAPEADO, falta __*validar*__ documentação - Responsável: __*Arthur*__\n- **/profile** ➜ MAPEADO, falta __*completar*__ documentação - Responsável: __*Arthur*__\n- **/sector** ➜ MAPEADO, falta __*completar*__ documentação - Responsável: __*Arthur*__\n- **/project** ➜ MAPEADO, falta __*completar*__ documentação - Responsável: __*Arthur*__\n- **/messages** ➜ MAPEADO, falta __*validar*__ documentação - Responsável: __*José*__\n- **/chat** ➜ MAPEADO, falta __*validar*__ documentação (não é prioritário) - Responsável: __*José*__",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearer": {
          "format": "JWT",
          "type": "http"
        }
      }
    }
  },
  "paths": {
...
    "/v1/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retorna informações do usuário autenticado.",
        "description": "Retorna informações do usuário autenticado, utilizando o token de acesso fornecido no header.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Informações do usuário autenticado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "4b2ca562-7b37-4b49-8ae3-17a8b9f1349a"
                    },
                    "cpf": {
                      "type": "string",
                      "example": "06093240479"
                    },
                    "name": {
                      "type": "string",
                      "example": "John Doe"
                    },
                    "email": {
                      "type": "string",
                      "example": "adminDaSilva@admin.com"
                    },
                    "phone": {
                      "type": "string",
                      "example": "1234567890"
                    },
                    "isActive": {
                      "type": "boolean",
                      "example": true
                    },
                    "profile": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "7e3f6c24-057f-4661-ab28-c8c9f72b5924"
                        },
                        "name": {
                          "type": "string",
                          "example": "default_user"
                        },
                        "description": {
                          "type": "string",
                          "example": "Permissão de usuário padrão do sistema"
                        }
                      }
                    },
                    "sector": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falha na autenticação. Isso pode ocorrer se o token não for fornecido ou se o token for inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Token not provided."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/list": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retorna uma PAGINADA lista de usuários.",
        "description": "Retorna uma lista de usuários cadastrados no sistema.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários.",
            "content": {
              "application/json": {
                "example": {
                  "listUser": [
                    {
                      "id": "7ea5f441-47bb-4eb5-8869-e69e10bbd167",
                      "cpf": "123456789",
                      "name": "Admin",
                      "email": "adminDaSilva@admin.com",
                      "phone": null,
                      "isActive": true,
                      "profile": {
                        "id": "e6f28377-5ff0-4bde-a471-c1d6d53cdd4c",
                        "name": "admin",
                        "description": "Administrador do sistema"
                      },
                      "sector": null
                    }
                  ],
                  "pagination": {
                    "path": "/users",
                    "page": 1,
                    "prev_page_url": false,
                    "next_page_url": false,
                    "total": 1
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falha na autenticação. Isso pode ocorrer se o token não for fornecido ou se o token for inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Token not provided."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Falha na autenticação. Usuário não tem premissão de acesso a rota.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Access denied."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/listAll": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retorna uma lista de TODOS os usuários.",
        "description": "Retorna uma lista de usuários cadastrados no sistema.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários.",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "7ea5f441-47bb-4eb5-8869-e69e10bbd167",
                    "cpf": "123456789",
                    "name": "Admin",
                    "email": "adminDaSilva@admin.com",
                    "phone": null,
                    "isActive": true,
                    "profile": {
                      "id": "e6f28377-5ff0-4bde-a471-c1d6d53cdd4c",
                      "name": "admin",
                      "description": "Administrador do sistema"
                    },
                    "sector": null
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Falha na autenticação. Isso pode ocorrer se o token não for fornecido ou se o token for inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Token not provided."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Falha na autenticação. Usuário não tem premissão de acesso a rota.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Access denied."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/{id}/update": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Atualiza um usuário.",
        "description": "Atualiza um usuário no sistema.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "4b2ca562-7b37-4b49-8ae3-17a8b9f1349a"
            },
            "description": "ID do usuário a ser atualizado."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cpf": {
                    "type": "string",
                    "example": "06093240479",
                    "description": "CPF do usuário, contendo 11 dígitos numéricos.",
                    "pattern": "^[0-9]{11}$"
                  },
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "adminDaSilva@admin.com",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string",
                    "example": "1234567890",
                    "description": "Número de telefone do usuário, sem formatação específica."
                  },
                  "birthDate": {
                    "type": "string",
                    "example": "1990-01-01",
                    "format": "date",
                    "description": "Data de nascimento do usuário, opcional."
                  },
                  "profile": {
                    "type": "string",
                    "example": "7e3f6c24-057f-4661-ab28-c8c9f72b5924",
                    "description": "UUID do perfil do usuário, opcional."
                  },
                  "sector": {
                    "type": "string",
                    "example": "7e3f6c24-057f-4661-ab28-c8c9f72b5924",
                    "description": "UUID do setor do usuário, opcional."
                  },
                  "isActive": {
                    "type": "boolean",
                    "example": true,
                    "description": "Status ativo do usuário, opcional."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário atualizado com sucesso."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação dos dados de entrada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "birthDate is invalid. Use  YYYY-MM-DD"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falha na autenticação. Isso pode ocorrer se o token não for fornecido ou se o token for inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Token not provided."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Falha na autenticação. Usuário não tem premissão de acesso a rota.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Access denied."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/{id}/update/profile": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Atualiza O PERFIL de UM usuário.",
        "description": "Atualiza O PERFIL de UM usuário no sistema.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "4b2ca562-7b37-4b49-8ae3-17a8b9f1349a"
            },
            "description": "ID do usuário a ser atualizado."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profileId": {
                    "type": "string",
                    "example": "7e3f6c24-057f-4661-ab28-c8c9f72b5924",
                    "description": "UUID do perfil do usuário, opcional."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil do usuário atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Perfil do usuário atualizado com sucesso."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falha na autenticação. Isso pode ocorrer se o token não for fornecido ou se o token for inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Token not provided."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Falha na autenticação. Usuário não tem premissão de acesso a rota.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 403
                    },
                    "message": {
                      "type": "string",
                      "example": "Access denied."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Perfil não encontrado."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/grant/post": {
      "post": {
        "tags": [
          "Grants"
        ],
        "summary": "Cria uma nova grant",
        "description": "Cria uma nova grant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "O nome da grant."
                  },
                  "note": {
                    "type": "string",
                    "description": "Uma nota ou descrição para a grant."
                  },
                  "route": {
                    "type": "string",
                    "description": "A rota associada à grant."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1v1~1grant~1post/post/requestBody/content/application~1json/schema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/401"
          },
          "403": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/403"
          },
          "500": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/500"
          }
        }
      }
    },
...
    "/v1/grant/getAll": {
      "get": {
        "tags": [
          "Grants"
        ],
        "summary": "Obtém todas as grants",
        "description": "Obtém todas as grants",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/paths/~1v1~1grant~1put~1%7Bid%7D/put/responses/200/content/application~1json/schema"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/401"
          },
          "403": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/403"
          },
          "404": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/404"
          },
          "500": {
            "$ref": "#/paths/~1v1~1chat~1post/post/responses/500"
          }
        }
      }
    },
...
  }
}